Address PR review on experiment 009
- move distinctive-feature ranking into its own module
- split vowel and consonant feature distances with vowel_ratio
- rerun experiment 009 and update leaderboard
