+++ src/soramimi_phonetic_search_dataset/__init__.py
@@ -7,9 +7,9 @@ from .dataset import (
     load_default_dataset,
     load_phonetic_search_dataset,
 )
+from .distinctive_feature import rank_by_distinctive_feature_distance
 from .evaluate import evaluate_ranking_function, evaluate_ranking_function_with_details
 from .ranking import (
-    rank_by_distinctive_feature_distance,
     rank_by_kanasim,
     rank_by_mora_editdistance,
     rank_by_phoneme_editdistance,
